requirements.txt
amqp==2.3.2
appnope==0.1.0
argh==0.26.2
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.4
bleach==3.1.0
boto==2.49.0
boto3==1.9.4
botocore==1.12.215
celery==4.2.1
certifi==2019.9.11
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
contextlib2==0.5.5
cookies==2.2.1
coverage==4.4.1
cryptography==2.3.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11.17
django-crispy-forms==1.6.0
django-environ==0.4.5
django-filter==2.1.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-redis==4.9.0
django-restframework==0.0.1
django-test-plus==1.0.18
djangorestframework==3.8.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
environ==1.0
environs==4.0.0
factory-boy==2.8.1
Faker==2.0.0
flake8==3.3.0
Flask==1.0.3
geojson==2.5.0
geojson-utils==0.0.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
humanize==0.5.1
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jwt==0.5.4
kombu==4.2.1
lazy-object-proxy==1.4.1
locustio==0.11.0
Markdown==3.1.1
markdown2==2.3.8
MarkupSafe==1.1.1
marshmallow==3.3.0
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.15.3
more-itertools==7.0.0
msgpack==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.4.0
pathtools==0.1.2
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.18
pyflakes==1.5.0
Pygments==2.3.1
pylint==1.7.2
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pymongo==3.7.1
pynamodb==4.2.0
pyrsistent==0.14.11
pytesseract==0.3.2
pytest==3.7.4
pytest-django==3.1.2
pytest-sugar==0.8.0
pytest-watch==4.1.0
python-dateutil==2.8.0
python-dotenv==0.10.5
pytz==2018.5
PyYAML==5.1.2
pyzmq==18.0.1
qtconsole==4.4.3
redis==3.3.11
requests==2.22.0
responses==0.5.1
s3transfer==0.1.13
schema==0.7.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.4.0
tzlocal==1.5.1
urllib3==1.25.3
vine==1.1.4
virtualenv==16.4.3
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==4.0
widgetsnbextension==3.4.2
wrapt==1.11.2
